Move GFX9 plot tests to dedicated module
