+++ projects/rocprofiler-compute/tests/test_mem_chart.py
@@ -18,13 +18,6 @@ from utils import mem_chart_gfx9, mem_chart_gfx11
 
 DEFAULT_TITLE = "3. Memory Chart (Normalization: per_kernel)"
 
-
-def render_gfx9_chart(metrics):
-    return common.strip_ansi(
-        mem_chart_gfx9.plot_mem_chart(metrics, chart_title=DEFAULT_TITLE)
-    )
-
-
 # =============================================================================
 # Tests for format_bw_human_readable function (gfx11)
 # =============================================================================
@@ -559,254 +552,6 @@ class TestIntegrationGfx11:
         # Should contain multiple unit types
         assert "TB/s" in chart
         assert "GB/s" in chart
-
-
-# =============================================================================
-# Tests for plot_mem_chart function (gfx9)
-# =============================================================================
-
-
-class TestPlotMemChartGfx9:
-    """Tests for gfx9 plot_mem_chart - CDNA memory chart generation."""
-
-    def test_full_sample_metrics_render_without_na_placeholders(self):
-        """Render complete gfx9 metrics without N/A placeholders."""
-        output = render_gfx9_chart(common.GFX9_SAMPLE_METRICS)
-        assert isinstance(output, str)
-        assert len(output) > 0
-        assert "N/A" not in output
-
-    def test_empty_metrics_render_expected_placeholders(self):
-        """Render 54 N/A placeholders when no gfx9 metrics are provided."""
-        output = render_gfx9_chart({})
-        assert isinstance(output, str)
-        assert len(output) > 0
-
-        # The renderer consumes 55 keys but emits 54 uppercase placeholders because
-        # missing Active CUs/Num CUs render together as "n/a: N/A".
-        assert output.count("N/A") == 54
-
-    def test_partial_metrics_render_values_and_placeholders(self):
-        """Render supplied gfx9 values and placeholders for missing metrics."""
-        partial = {"HBM Rd": 100}
-        output = render_gfx9_chart(partial)
-        assert isinstance(output, str)
-        assert len(output) > 0
-        assert "Rd:  100" in output
-        assert "N/A" in output
-
-    def test_contains_complete_cdna_architecture(self):
-        """CDNA output contains every component enabled by the renderer."""
-        output = render_gfx9_chart(common.GFX9_SAMPLE_METRICS)
-        expected_components = (
-            "Instr Buff",
-            "Instr Dispatch",
-            "Exec",
-            "LDS",
-            "Vector L1 Cache",
-            "Scalar L1D Cache",
-            "Instr L1 Cache",
-            "L2 Cache",
-            "xGMI/PCIe",
-            "Fabric",
-            "HBM",
-        )
-
-        for component in expected_components:
-            assert component in output, f"Missing CDNA component: {component}"
-
-        assert re.search(r"(?<!x)GMI(?!/PCIe)", output)
-
-    def test_gfx9_heading_outside_chart_and_directional_connectors(self):
-        """CDNA output prints the heading once outside the chart, with connectors."""
-        output = render_gfx9_chart(common.GFX9_SAMPLE_METRICS)
-        output_lines = output.splitlines()
-
-        assert output_lines[0] == DEFAULT_TITLE
-        assert "3. Memory Chart" not in "\n".join(output_lines[1:])
-        assert "Instr Buff" in output_lines[4]
-        assert output.count("(Normalization: per_kernel)") == 1
-
-        for connector_label in ("Req", "Fetch", "Rd", "Wr", "Atomic"):
-            assert connector_label in output
-
-        assert re.search(r"<(?!-+>)-{3,}", output)
-        assert re.search(r"(?<![<-])-{3,}>", output)
-        assert re.search(r"<-{3,}>", output)
-
-    @pytest.mark.parametrize(
-        (
-            "metric_name",
-            "metric_value",
-            "expected_text",
-            "expected_row_index",
-            "expected_column_slice",
-        ),
-        [
-            pytest.param(
-                "SALU",
-                7101,
-                "SALU: 7101",
-                6,
-                slice(38, 58),
-                id="salu-instruction",
-            ),
-            pytest.param(
-                "SMEM",
-                7102,
-                "SMEM: 7102",
-                9,
-                slice(38, 58),
-                id="smem-instruction",
-            ),
-            pytest.param(
-                "VGPR",
-                7201,
-                "RVGPRseq:  7201",
-                11,
-                slice(58, 78),
-                id="vgpr-allocation",
-            ),
-            pytest.param(
-                "SGPR",
-                7202,
-                "SGPRs:  7202",
-                14,
-                slice(58, 78),
-                id="sgpr-allocation",
-            ),
-            pytest.param(
-                "Wavefronts",
-                7301,
-                "7301",
-                28,
-                slice(58, 78),
-                id="wavefront-count",
-            ),
-            pytest.param(
-                "Workgroups",
-                7302,
-                "7302",
-                32,
-                slice(58, 78),
-                id="workgroup-count",
-            ),
-            pytest.param(
-                "LDS Req",
-                1234,
-                "Req: 1234",
-                7,
-                slice(78, 95),
-                id="lds-request",
-            ),
-            pytest.param(
-                "LDS Latency",
-                321,
-                "Lat:    321 cycles",
-                9,
-                slice(95, 122),
-                id="lds-latency",
-            ),
-            pytest.param(
-                "L2 Hit",
-                87,
-                "Hit:     87 %",
-                9,
-                slice(140, 165),
-                id="l2-hit-rate",
-            ),
-            pytest.param(
-                "L2 Rd Lat",
-                2468,
-                "Rd:   2468",
-                27,
-                slice(140, 165),
-                id="l2-read-latency",
-            ),
-            pytest.param(
-                "L2 Wr Lat",
-                1357,
-                "Wr:   1357",
-                29,
-                slice(140, 165),
-                id="l2-write-latency",
-            ),
-            pytest.param(
-                "VL1 Rd",
-                7401,
-                "Rd: 7401",
-                15,
-                slice(78, 95),
-                id="vector-l1-read",
-            ),
-            pytest.param(
-                "VL1 Wr",
-                7402,
-                "Wr: 7402",
-                17,
-                slice(78, 95),
-                id="vector-l1-write",
-            ),
-            pytest.param(
-                "VL1 Hit",
-                2581,
-                "Hit:   2581 %",
-                15,
-                slice(95, 122),
-                id="vector-l1-hit-rate",
-            ),
-            pytest.param(
-                "sL1D Lat",
-                3692,
-                "Lat:   3692 cycles",
-                29,
-                slice(95, 122),
-                id="scalar-l1d-latency",
-            ),
-            pytest.param(
-                "IL1 Hit",
-                4703,
-                "Hit:   4703 %",
-                35,
-                slice(95, 122),
-                id="instruction-l1-hit-rate",
-            ),
-            pytest.param(
-                "Fabric Rd Lat",
-                5814,
-                "Rd:   5814",
-                20,
-                slice(183, 207),
-                id="fabric-read-latency",
-            ),
-            pytest.param(
-                "HBM Rd",
-                6925,
-                "Rd: 6925",
-                18,
-                slice(207, 221),
-                id="hbm-read-bandwidth",
-            ),
-        ],
-    )
-    def test_metric_routes_to_expected_chart_region(
-        self,
-        metric_name,
-        metric_value,
-        expected_text,
-        expected_row_index,
-        expected_column_slice,
-    ):
-        """Place each gfx9 metric in the expected chart region."""
-        output_lines = render_gfx9_chart({metric_name: metric_value}).splitlines()
-
-        assert expected_text in output_lines[expected_row_index][expected_column_slice]
-
-    def test_empty_placeholders_do_not_render_metric_suffixes(self):
-        """Omit percentage and cycle suffixes from N/A placeholders."""
-        output = render_gfx9_chart({})
-
-        assert re.search(r"N/A[ \t]*(?:%|cycles)", output) is None
 
 
 @pytest.mark.parametrize(
